working, but no exit animations...
