+++ package.json
@@ -17,6 +17,7 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
+    "@dnd-kit/core": "^6.0.5",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "@productboard/slate-edit-list": "^0.21.5",
