+++ src/components/ItemList/ListContext.js
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+const ListContext = createContext({
+  items: [],
+  columns: 0,
+});
+
+export default ListContext;
