+++ src/styles/components/list/_item-list.scss
@@ -1,12 +1,12 @@
 .item-list {
   flex: 1;
   display: flex;
+  position: relative;
 
   &__container,
   &__sizer {
     flex: 1;
     display: flex;
-    position: relative;
   }
 
   &__grid {
@@ -20,12 +20,34 @@
     width: 100%;
   }
 
-  &--empty {}
+  &__empty {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 1;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
 
-  &__emptyMessage {
-    flex: 1;
-    align-self: center;
-    text-align: center;
+  &__drop-overlay {
+    background: var(--transparent-dark);
+    color: white;
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 1;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    &--active {
+      background: var(--background);
+    }
   }
 }
