+++ stories/ItemList.stories.js
@@ -1,12 +1,31 @@
 import React from 'react';
 import faker from 'faker';
 import { v4 as uuid } from 'uuid';
-import ItemList from '../src/components/List/ItemList';
+import ItemList from '../src/components/ItemList/ItemList';
 import '../src/styles/_all.scss';
 import Node from '../src/components/Node';
 import SessionCard from '../src/components/Cards/SessionCard';
 import ProtocolCard from '../src/components/Cards/ProtocolCard';
 import DataCard from '../src/components/Cards/DataCard';
+import { DndContext, useDraggable } from '@dnd-kit/core';
+
+function Draggable(props) {
+  const { attributes, listeners, setNodeRef, transform } = useDraggable({
+    id: 'draggable',
+    data: {
+      type: 'draggable',
+    },
+  });
+  const style = transform ? {
+    transform: `translate3d(${transform.x}px, ${transform.y}px, 0)`,
+  } : undefined;
+
+  return (
+    <button ref={setNodeRef} style={style} {...listeners} {...attributes}>
+      {props.children}
+    </button>
+  );
+}
 
 const TestCard = (attributes) => (
   <div
@@ -50,7 +69,7 @@ const mockItems = (length = 100) => [...Array(length)].map(() => (
       updatedAt: faker.date.recent().toUTCString(),
     },
   }
-));
+)).sort((item1, item2) => item1.attributes.name > item2.attributes.name);
 
 export default {
   title: 'Components/ItemList',
@@ -62,6 +81,7 @@ export default {
         1000: mockItems(1000),
         100: mockItems(100),
         10: mockItems(10),
+        0: [],
       },
       control: { type: 'radio' },
     },
@@ -89,25 +109,28 @@ export default {
 
 const Template = (args) => (
   <>
-    <div
-      style={{
-        display: 'flex',
-        height: '400px',
-        width: '100%',
-        border: '1px solid tomato',
-        '--base-font-size': '12px',
-        resize: 'both',
-        overflow: 'auto',
-      }}
-    >
-      <ItemList
-        {...args}
-        cardColumnBreakpoints={{
-          800: 2,
-          1200: 3,
+    <DndContext>
+      <Draggable>Test</Draggable>
+      <div
+        style={{
+          display: 'flex',
+          height: '400px',
+          width: '100%',
+          border: '1px solid tomato',
+          '--base-font-size': '12px',
+          resize: 'both',
+          overflow: 'auto',
         }}
-      />
-    </div>
+      >
+        <ItemList
+          {...args}
+          cardColumnBreakpoints={{
+            800: 2,
+            1200: 3,
+          }}
+        />
+      </div>
+    </DndContext>
   </>
 );
